minor changes to data, relationships. deleted unused models.

+++ app/models/Grade.java
@@ -8,7 +8,7 @@ import play.db.jpa.*;
 @Entity
 public class Grade extends Model {
 
-	public int grade;
+	public double grade;
 
 	@ManyToOne
 	public Activity assignment;
